Speedup waiting for cluster states in some ML tests
Found this trying to reproduce a failure. The ML integ-tests spend a
massive amount of time waiting on cluster states. Doing so via a busy
wait that fetches the state from a random node is rather CPU intensive
and due to backoff in the busy assert, will waste seconds per run.
-> moved the CS wait logic from snapshot tests to the top level and made
use of it in some ML tests, saving both CPU resources and 10s+ of time
per run of the integTests on my Macbook.
