+++ x-pack/plugin/ml/src/internalClusterTest/java/org/elasticsearch/xpack/ml/integration/MlDistributedFailureIT.java
@@ -28,7 +28,6 @@ import org.elasticsearch.index.query.TermsQueryBuilder;
 import org.elasticsearch.persistent.PersistentTaskResponse;
 import org.elasticsearch.persistent.PersistentTasksClusterService;
 import org.elasticsearch.persistent.PersistentTasksCustomMetadata;
-import org.elasticsearch.persistent.PersistentTasksCustomMetadata.PersistentTask;
 import org.elasticsearch.persistent.UpdatePersistentTaskStatusAction;
 import org.elasticsearch.xcontent.ToXContent;
 import org.elasticsearch.xcontent.XContentBuilder;
@@ -719,15 +718,19 @@ public class MlDistributedFailureIT extends BaseMlIntegTestCase {
         // https://github.com/elastic/elasticsearch/pull/50907 - now that the cluster state is stored
         // in a Lucene index it can take a while to update when there are many updates in quick
         // succession, like we see in internal cluster tests of node failure scenarios
-        assertBusy(() -> {
-            ClusterState clusterState = clusterAdmin().prepareState().get().getState();
-            List<PersistentTask<?>> tasks = findTasks(clusterState, Set.of(DATAFEED_TASK_NAME, JOB_TASK_NAME));
-            assertNotNull(tasks);
-            assertEquals("Expected 2 tasks, but got [" + tasks + "]", 2, tasks.size());
-            for (PersistentTask<?> task : tasks) {
-                assertFalse(needsReassignment(task.getAssignment(), clusterState.nodes()));
+        awaitClusterState(state -> {
+            List<PersistentTasksCustomMetadata.PersistentTask<?>> tasks = findTasks(state, Set.of(DATAFEED_TASK_NAME, JOB_TASK_NAME));
+            if (tasks == null || tasks.size() != 2) {
+                return false;
             }
-
+            for (PersistentTasksCustomMetadata.PersistentTask<?> task : tasks) {
+                if (needsReassignment(task.getAssignment(), state.nodes())) {
+                    return false;
+                }
+            }
+            return true;
+        });
+        assertBusy(() -> {
             GetJobsStatsAction.Request jobStatsRequest = new GetJobsStatsAction.Request(jobId);
             JobStats jobStats = client().execute(GetJobsStatsAction.INSTANCE, jobStatsRequest).actionGet().getResponse().results().get(0);
             assertEquals(JobState.OPENED, jobStats.getState());
